refactor log_if to better divide interfaces and add a prelude for glob imports
